v 0.9.9.0.2 Big Database Changes and bugfixes
- The MongoDB design has been reworked. It still needs some tweaking
here and there (coming in next updates).
Now also supports storing location of media files and automatically
download them to NovA (for sftp/http links) The documentation will be
adjusted soon.
- Added Confidences for Annotations. These will for now be stored only
when using a database.
- Now jumps to annotation when selected in list
- Bugfix: When the rangeslider was too small,  it didnt follow the media
anymore (if option was selected)
- Added Metro Theme. For now in a seperate solution/binary. A more
modern look and feel.
-Bugfix: Fixed timetrack labels (sometimes there were overlaps)
- Bugfix: When tracks are exported as sampled list, the table headers
are now again the tier names.
-Export Annotracks as SVG or PNG