Database adjustments
1. Annotations are now shown in a listview (no more #)
(Other elements will be handled the same way)
2. Annotype and Role now have the attribute isValid. Instead of
deleting, they are kept in the database but marked as invalid. This
eases reference errors. (Later a clean database function can check all
invalid entries and clean the db if they are not needed anymore).. Other
collections will get this entry as well..